requirements.txt
Flask
Flask-Cors
requests
openai
python-dotenv
langchain
